move asset components under assets dir
